fix(core-network): handle decimal numbers in menu API responses
서버가 가격/금액을 소수(4500.00)로 반환하여 Int 파싱 실패.
MenuDto, MenuDetailDto, RecipeDto의 숫자 필드를 Double로 변경하고
매퍼에서 toInt() 변환

+++ core/core-data/src/main/kotlin/com/team/chord/core/data/datasource/remote/RemoteMenuDataSource.kt
@@ -46,7 +46,7 @@ class RemoteMenuDataSource @Inject constructor(
 
     override suspend fun getMenuRecipes(menuId: Long): Pair<List<MenuRecipe>, Int> {
         val result = safeApiCall { menuApi.getMenuRecipes(menuId) }
-        return result.recipes.map { it.toDomain() } to result.totalCost
+        return result.recipes.map { it.toDomain() } to result.totalCost.toInt()
     }
 
     override suspend fun checkMenuDuplicate(menuName: String, ingredientNames: List<String>?): CheckDupResult {
