+++ core/core-network/src/main/kotlin/com/team/chord/core/network/dto/menu/MenuDtos.kt
@@ -36,7 +36,7 @@ data class RecipeTemplateDto(
 data class MenuDto(
     val menuId: Long,
     val menuName: String,
-    val sellingPrice: Int,
+    val sellingPrice: Double,
     val costRate: Float,
     val marginGradeCode: String,
     val marginGradeName: String,
@@ -48,15 +48,15 @@ data class MenuDetailDto(
     val menuId: Long,
     val menuName: String,
     val workTime: Int,
-    val sellingPrice: Int,
+    val sellingPrice: Double,
     val marginRate: Float,
-    val totalCost: Int,
+    val totalCost: Double,
     val costRate: Float,
-    val contributionMargin: Int,
+    val contributionMargin: Double,
     val marginGradeCode: String,
     val marginGradeName: String,
     val marginGradeMessage: String,
-    val recommendedPrice: Int? = null,
+    val recommendedPrice: Double? = null,
 )
 
 @Serializable
@@ -65,15 +65,15 @@ data class RecipeDto(
     val menuId: Long,
     val ingredientId: Long,
     val ingredientName: String,
-    val amount: Int,
+    val amount: Double,
     val unitCode: String,
-    val price: Int,
+    val price: Double,
 )
 
 @Serializable
 data class RecipeListDto(
     val recipes: List<RecipeDto>,
-    val totalCost: Int,
+    val totalCost: Double,
 )
 
 @Serializable
